requirements.txt
argparse
colorlog
openai
python-dotenv